% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/pedigree.R
\name{editPed}
\alias{editPed}
\title{Edits a disordered or incomplete pedigree,
   1_ add labels for the sires and dams not listed as labels before.
   2_ order pedigree based on recursive calls to getGenAncestors.}
\usage{
editPed(sire, dam, label, verbose = FALSE)
}
\arguments{
\item{sire}{integer vector or factor representation of the sires}

\item{dam}{integer vector or factor representation of the dams}

\item{label}{character vector of labels}

\item{verbose}{logical to print the row of the pedigree that the
function is ordering. Default is FALSE.}
}
\value{
a data frame with the pedigree ordered.
}
\description{
Edits a disordered or incomplete pedigree,
   1_ add labels for the sires and dams not listed as labels before.
   2_ order pedigree based on recursive calls to getGenAncestors.
}

