% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{expandPedigreeSelfing}
\alias{expandPedigreeSelfing}
\title{Expand Pedigree to Account a pedigree object and expands it to account for selfing generations.}
\usage{
expandPedigreeSelfing(ped, sepChar = "-F", verbose = FALSE)
}
\arguments{
\item{ped}{An object of class "pedigree"}

\item{sepChar}{Character used for expanded pedigree IDs (default: '-F')}

\item{verbose}{Logical, whether to print progress (default: FALSE)}
}
\value{
A new pedigree object with expanded entries for selfing generations
}
\description{
Expand Pedigree to Account a pedigree object and expands it to account for selfing generations.
}
