% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{getAInv}
\alias{getAInv}
\title{Inverse of the additive relationship matrix}
\usage{
getAInv(ped, gamma = NULL)
}
\arguments{
\item{ped}{\code{\link{pedigree}}}

\item{gamma}{single number or matrix descring relationship between metafounder}
}
\value{
matrix (\link[Matrix]{dsCMatrix-class} - symmetric sparse)
}
\description{
Returns the inverse of additive relationship matrix for the
  pedigree.
}
\examples{
ped <- pedigree(sire = c(NA, NA, 1,  1, 4, 5),
                dam =  c(NA, NA, 2, NA, 3, 2),
                label = 1:6)
(AInv <- getAInv(ped))

}
