\name{getASelfing}
\alias{getASelfing}
\title{Additive Relationship Matrix taking into account selfig}
\usage{
getASelfing(ID,Par1,Par2,nCycles,fileNewPed=NULL)
}
\arguments{
\item{ID}{is a vector with the individual IDs.}
\item{Par1}{is a vector with the IDs of one of the parents.}
\item{Par2}{is a vector with the IDs of the other parent.}
\item{nCycles}{is a vector that indicates the number of selfing cycles for each individual.}
\item{fileNewPed}{Output csv file (comma separated value) with columns ‘label’, ‘sire’, ‘dam’, 
      with the full pull pedigree expanded taking into account the selfing cycles}
}
\value{

}
\description{

}
\examples{

}
\keyword{array}
