% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{getGeneration}
\alias{getGeneration}
\title{Get or Calculate Generation Numbers for a Pedigree}
\usage{
getGeneration(ped)
}
\arguments{
\item{ped}{An object of class "pedigree"}
}
\value{
The updated pedigree object with the generation slot filled.
}
\description{
This function retrieves or calculates the generation numbers for a pedigree object.
If the generation information is not already stored in the pedigree object, it
calculates it and updates the pedigree object.
}
\details{
This function first checks if generation information is already stored in the
  pedigree object. If not, it calculates the generation numbers using `getGenAncestors`.
  The calculated generation numbers are then stored in the pedigree object's generation slot.
}
\examples{
ped <- pedigree(sire = c(NA, NA, 1, 1, 4, 5),
                dam  = c(NA, NA, 2, NA, 3, 2),
                label = 1:6)
ped <- getGeneration(ped)
print(ped@generation)

}
\seealso{
\code{\link{getGenAncestors}} for the underlying function used to calculate
  generation numbers, \code{\link{ped2DF}} for converting a pedigree object to a data frame.
}
