% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{getT}
\alias{getT}
\title{Gene flow from a pedigree}
\usage{
getT(ped)
}
\arguments{
\item{ped}{\code{\link{pedigree}}}
}
\value{
matrix (\link[Matrix]{dtCMatrix-class} - lower unitriangular sparse)
}
\description{
Get gene flow matrix from a pedigree.
}
\examples{
ped <- pedigree(sire = c(NA, NA, 1,  1, 4, 5),
                dam =  c(NA, NA, 2, NA, 3, 2),
                label = 1:6)
(T <- getT(ped))

}
