% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{inbreeding}
\alias{inbreeding}
\title{Inbreeding coefficients from a pedigree}
\usage{
inbreeding(ped, gamma = NULL)
}
\arguments{
\item{ped}{\code{\link{pedigree}}}

\item{gamma}{single number or matrix descring relationship between metafounder}
}
\value{
the inbreeding coefficients as a numeric vector
}
\description{
Create the inbreeding coefficients for a pedigree without metafounders
             according to the algorithm given in "Comparison of four direct algorithms for computing inbreeding
             coefficients" by Mehdi Sargolzaei and Hiroaki Iwaisaki, Animal Science Journal (2005) 76, 401--406.
             The method used for pedigree with metafounders is based on the algorithm proposed
             by Legarra et al. (2015). Ancestral Relationships Using Metafounders: Finite Ancestral
             Populations and Across Population Relationships.
             Genetics. Jun;200(2):455-68. doi: 10.1534/genetics
}
