<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======

% Generated by roxygen2: do not edit by hand

% Generated by roxygen2 (4.0.2.9000): do not edit by hand

>>>>>>> e2765e1fc8f94c5565ab726465d35c811eb6dc30
% Please edit documentation in R/pedigree.R
\name{ped2DF}
\alias{ped2DF}
\title{Convert a pedigree to a data frame}
\usage{
ped2DF(x)
}
\arguments{
\item{x}{a pedigree object of class \linkS4class{pedigree}}
}
\value{
a data frame
}
\description{
Express a pedigree as a data frame with \code{sire} and
\code{dam} stored as factors.  If the pedigree is an object of
class \linkS4class{pedinbred} then the inbreeding coefficients are
appended as the variable \code{F}
}

