<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.0.2.9000): do not edit by hand
>>>>>>> dd83974afe0494cdbc55467f9ce98a9beb64c819
% Please edit documentation in R/AllClass.R
\docType{class}
\name{pedigree-class}
\alias{pedigree-class}
\title{Pedigree class}
\description{
Pedigree class
}

