% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeTools.R
\docType{package}
\name{pedigreeTools-package}
\alias{pedigreeTools}
\alias{pedigreeTools-package}
\title{pedigreeTools: Versatile Functions for Working with Pedigrees}
\description{
Tools to sort, edit and prune pedigrees and to extract the inbreeding
coefficients and the relationship matrix (includes code for pedigrees from
self-pollinated species). The use of pedigree data is central to genetics
research within the animal and plant breeding communities to predict breeding
values. The relationship matrix between the individuals can be derived from
pedigree structure following the algorithms described for example in
Vazquez et al., 2010 <doi:10.2527/jas.2009-1952>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Rpedigree/pedigreeTools/}
}

}
\keyword{internal}
