% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Dmat}
\alias{Dmat}
\title{Diagonal of D in the A = TDT' factorization.}
\usage{
Dmat(ped)
}
\arguments{
\item{ped}{an object that inherits from class \linkS4class{pedigree}}
}
\value{
a numeric vector
}
\description{
Determine the diagonal factor in the decomposition of the
relationship matrix A as TDT' where T is unit lower triangular.
}

