
% Generated by roxygen2: do not edit by hand

% Generated by roxygen2 (4.0.2.9000): do not edit by hand

% Please edit documentation in R/pedigree.R
\name{Dmat}
\alias{Dmat}
\title{Diagonal of D in the A = TDT' factorization.}
\usage{
Dmat(ped)
}
\arguments{
\item{ped}{an object that inherits from class \linkS4class{pedigree}}
}
\value{
a numeric vector
}
\description{
Determine the diagonal factor in the decomposition of the
relationship matrix A as TDT' where T is unit lower triangular.
}
\examples{
ped <- pedigree(sire=c(NA,NA,1,1,4,5), dam=c(NA,NA,2,NA,3,2), label=1:6)
Dmat(ped)
}

