% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{Dmat}
\alias{Dmat}
\alias{getD}
\alias{DInvmat}
\title{Mendelian sampling variance}
\usage{
Dmat(ped, gamma = NULL, vector = TRUE)

getD(ped, gamma = NULL, vector = TRUE)

DInvmat(ped, gamma = NULL, vector = TRUE)
}
\arguments{
\item{ped}{\code{\link{pedigree}}}

\item{gamma}{TODO}

\item{vector}{logical, return a vector or sparse matrix}
}
\value{
a numeric vector
}
\description{
Determine the diagonal factor in the decomposition of the
  relationship matrix A as TDT' where T is unit lower triangular.
}
\section{Functions}{
\itemize{
\item \code{getD()}: Mendelian sampling variance

\item \code{DInvmat()}: Mendelian sampling precision (= 1 / variance)

}}
\examples{
ped <- pedigree(sire = c(NA, NA, 1,  1, 4, 5),
                dam =  c(NA, NA, 2, NA, 3, 2),
                label = 1:6)
(D <- getD(ped))
(DInv <- getDInv(ped))

}
