% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getA}
\alias{getA}
\title{Additive Relationship Matrix}
\usage{
getA(ped)
}
\arguments{
\item{ped}{a pedigree that includes the individuals who occur in svec
which to restrict the relationship matrix. If \code{labs} is a
factor then the levels of the factor are used as the labels.
Default is the complete set of labels in the pedigree.}
}
\value{
an object that inherits from \linkS4class{CHMfactor}
}
\description{
Additive Relationship Matrix
}

