% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{getA}
\alias{getA}
\title{Additive Relationship Matrix}
\usage{
getA(ped)
}
\arguments{
\item{ped}{a pedigree that includes the individuals who occur in svec
which to restrict the relationship matrix. If \code{labs} is a
factor then the levels of the factor are used as the labels.
Default is the complete set of labels in the pedigree.}
}
\value{
an object that inherits from \linkS4class{CHMfactor}
}
\description{
Returns the additive relationship matrix for the pedigree \code{ped}.
}
\examples{
ped <- pedigree(sire=c(NA,NA,1,1,4,5), dam=c(NA,NA,2,NA,3,2), label=1:6)
getA(ped)
}
\keyword{array}

