% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{getA}
\alias{getA}
\title{Additive relationship matrix}
\usage{
getA(ped, labs = NULL)
}
\arguments{
\item{ped}{\code{\link{pedigree}}}

\item{labs}{a character vector or a factor giving individual labels to
which to restrict the relationship matrix and corresponding factor. If
\code{labs} is a factor then the levels of the factor are used as the
labels. Default is the complete set of individuals in the pedigree.}
}
\value{
matrix (\link[Matrix]{dsCMatrix-class} - symmetric sparse)
}
\description{
Returns the additive relationship matrix for the pedigree.
}
\examples{
ped <- pedigree(sire = c(NA, NA, 1,  1, 4, 5),
                dam =  c(NA, NA, 2, NA, 3, 2),
                label = 1:6)
(A <- getA(ped))

}
