
% Generated by roxygen2: do not edit by hand


% Generated by roxygen2: do not edit by hand

% Generated by roxygen2 (4.0.2.9000): do not edit by hand


% Please edit documentation in R/pedigree.R
\name{getAInv}
\alias{getAInv}
\title{Inverse of the Relationship Matrix}
\usage{
getAInv(ped)
}
\arguments{
\item{ped}{a pedigree that includes the individuals who occur in svec
which to restrict the relationship matrix. If \code{labs} is a
factor then the levels of the factor are used as the labels.
Default is the complete set of labels in the pedigree.}
}
\value{
an object that inherits from \linkS4class{CHMfactor}
}
\description{
Inverse of the Relationship Matrix
}
\examples{
ped <- pedigree(sire=c(NA,NA,1,1,4,5), dam=c(NA,NA,2,NA,3,2), label=1:6)
getAInv(ped)
}

