<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.0.2.9000): do not edit by hand
>>>>>>> dd83974afe0494cdbc55467f9ce98a9beb64c819
% Please edit documentation in R/pedigree.R
\name{getGenAncestors}
\alias{getGenAncestors}
\title{Counts number of generations of ancestors for one subject. Use recursion.}
\usage{
<<<<<<< HEAD
getGenAncestors(pede, id, ngen = NULL)
=======
getGenAncestors(pede, id)
>>>>>>> dd83974afe0494cdbc55467f9ce98a9beb64c819
}
\arguments{
\item{pede}{data frame with a pedigree and a column for the number of
generations of each subject.}

\item{id}{subject for which we want the number of generations.}
<<<<<<< HEAD

\item{ngen}{number of generation}
=======
>>>>>>> dd83974afe0494cdbc55467f9ce98a9beb64c819
}
\value{
a data frame object with the pedigree and generation of
  ancestors for subject id.
}
\description{
Counts number of generations of ancestors for one subject. Use recursion.
}

