<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======

% Generated by roxygen2: do not edit by hand

% Generated by roxygen2 (4.0.2.9000): do not edit by hand

>>>>>>> e2765e1fc8f94c5565ab726465d35c811eb6dc30
% Please edit documentation in R/pedigree.R
\name{getGenAncestors}
\alias{getGenAncestors}
\title{Counts number of generations of ancestors for one subject. Use recursion.}
\usage{
<<<<<<< HEAD
getGenAncestors(pede, id, ngen = NULL)
=======

getGenAncestors(pede, id, ngen = NULL)

getGenAncestors(pede, id)

>>>>>>> e2765e1fc8f94c5565ab726465d35c811eb6dc30
}
\arguments{
\item{pede}{data frame with a pedigree and a column for the number of
generations of each subject.}

\item{id}{subject for which we want the number of generations.}

<<<<<<< HEAD
\item{ngen}{number of generation}
=======

\item{ngen}{number of generation}


>>>>>>> e2765e1fc8f94c5565ab726465d35c811eb6dc30
}
\value{
a data frame object with the pedigree and generation of
  ancestors for subject id.
}
\description{
Counts number of generations of ancestors for one subject. Use recursion.
}

