
% Generated by roxygen2: do not edit by hand

% Generated by roxygen2 (4.0.2.9000): do not edit by hand

% Please edit documentation in R/pedigree.R
\name{getGenAncestors}
\alias{getGenAncestors}
\title{Counts number of generations of ancestors for one subject. Use recursion.}
\usage{

getGenAncestors(pede, id, ngen = NULL)

getGenAncestors(pede, id)

}
\arguments{
\item{pede}{data frame with a pedigree and a column for the number of
generations of each subject.}

\item{id}{subject for which we want the number of generations.}


\item{ngen}{number of generation}


}
\value{
a data frame object with the pedigree and generation of
  ancestors for subject id.
}
\description{
Counts number of generations of ancestors for one subject. Use recursion.
}

