% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ped2DF}
\alias{ped2DF}
\title{Convert a pedigree to a data frame}
\usage{
ped2DF(x)
}
\arguments{
\item{x}{a pedigree object of class \linkS4class{pedigree}}
}
\value{
a data frame
}
\description{
Express a pedigree as a data frame with \code{sire} and
\code{dam} stored as factors.  If the pedigree is an object of
class \linkS4class{pedinbred} then the inbreeding coefficients are
appended as the variable \code{F}
}

