<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======

% Generated by roxygen2: do not edit by hand

% Generated by roxygen2 (4.0.2.9000): do not edit by hand

>>>>>>> e2765e1fc8f94c5565ab726465d35c811eb6dc30
% Please edit documentation in R/AllClass.R
\docType{class}
\name{pedigree-class}
\alias{pedigree-class}
\title{Pedigree class}
\description{
Pedigree class
}

