% Generated by roxygen2: do not edit by hand
\name{prunePed}
\alias{prunePed}
\title{Subsets a pedigree for a specified vector of individuals upto a 
specified number of previous generations using Recursion.}
\usage{
prunePed(ped, selectVector, ngen = 2)
}
\arguments{
\item{ped}{Data Frame pedigree to be subset}

\item{selectVector}{Vector of individuals to select from pedigree}

\item{ngen}{Number of previous generations of parents to select starting from selectVector.}
}
\value{
Returns Subsetted pedigree as a DataFrame.
}
\description{
Subsets a pedigree for a specified vector of individuals upto a 
specified number of previous generations using Recursion.
}

